+++ tests/test_tile_image.py
@@ -1,6 +1,6 @@
 from tests.utils.tile import Tile
 from tests.utils.html import check_html
-from tests.conftest import assert_valid_html
+from tests.utils.html import assert_valid_html
 from tests import defaults
